ci: crio: make md2man install safer
If md2man is already installed, skip linking it again,
as if we are running on an already installed tree, the
destination may already be the same file, and the ln
then fails.

(we could rm and then re-install, or we could ignore
the ln fail - let's start with just skipping it)

+++ .ci/install_crio.sh
@@ -24,7 +24,12 @@ if [ ! -d "$GOBIN" ]
 then
         mkdir -p "$GOBIN"
 fi
-ln -sf $(command -v go-md2man) "$GOBIN"
+
+# What if it already exists - ignore it or remove it?
+#rm -f "${GOBIN}/md2man"
+if [ ! -e "${GOBIN}/go-md2man" ]; then
+	ln -sf $(command -v go-md2man) "$GOBIN"
+fi
 
 echo "Get CRI Tools"
 critools_repo="github.com/kubernetes-incubator/cri-tools"
